feat: reorganize js libraries category
