+++ apps/web/src/app/components/analysis-result/cards/minifier.tsx
@@ -24,5 +24,3 @@ export const MinifierCard: FC<{
     />
   );
 };
-
-export default MinifierCard;
