+++ apps/web/src/app/components/analysis-result/cards/modules.tsx
@@ -20,5 +20,3 @@ export const ModulesCard: FC<{
     />
   );
 };
-
-export default ModulesCard;
