+++ packages/analyzer/src/features/dates/detect.ts
@@ -1,0 +1,24 @@
+import { Browser, Page } from 'playwright';
+import { Resources } from '../../resources.js';
+import { patterns } from './patterns/index.js';
+import { calculateResults } from '../../calculate-results.js';
+
+export const detect = async (
+  page: Page,
+  browser: Browser,
+  resources: Resources
+) => {
+  const { result, getAllResultsWithConfidence } = await calculateResults(
+    resources,
+    page,
+    browser,
+    patterns
+  );
+
+  return {
+    type: 'dates',
+    name: result.name,
+    confidence: result.confidence,
+    secondaryMatches: getAllResultsWithConfidence(0.3, true),
+  };
+};
