+++ packages/analyzer/src/features/http-client/patterns/fetch.ts
