+++ packages/analyzer/src/features/http-client/patterns/got.ts
