+++ packages/analyzer/src/features/http-client/patterns/index.ts
