+++ packages/analyzer/src/features/http-client/patterns/superagent.ts
