+++ packages/analyzer/src/features/index.ts
@@ -4,7 +4,11 @@ export { detect as detectMinifier } from './minifier/detect.js';
 export { detect as detectStylingProcessor } from './styling-processor/detect.js';
 export { detect as detectModules } from './module/detect.js';
 export { detect as detectUILibrary } from './ui-library/detect.js';
-export { detect as detectJSLibraries } from './js-libraries/detect.js';
+export { detect as detectStateManagement } from './state-management/detect.js';
+export { detect as detectHttpClient } from './http-client/detect.js';
+export { detect as detectRouter } from './router/detect.js';
+export { detect as detectTranslations } from './translations/detect.js';
+export { detect as detectDatesLibrary } from './dates/detect.js';
 export { detect as detectStylingLibraries } from './styling-libraries/detect.js';
 export { detect as detectTranspiler } from './transpiler/detect.js';
 export { getStats } from './getStats.js';
@@ -12,10 +16,14 @@ export { getStats } from './getStats.js';
 export { patterns as bundlerPatterns } from './bundler/patterns/index.js';
 export { patterns as frameworkPatterns } from './framework/patterns/index.js';
 export { patterns as minifierPatterns } from './minifier/patterns/index.js';
+export { patterns as routerPatterns } from './router/patterns/index.js';
+export { patterns as translationsPatterns } from './translations/patterns/index.js';
+export { patterns as stateManagementPatterns } from './state-management/patterns/index.js';
+export { patterns as httpClientPatterns } from './http-client/patterns/index.js';
 export { cssInJsPatterns } from './styling-processor/patterns/css-in-js/index.js';
 export { preprocessorPatterns } from './styling-processor/patterns/preprocessor/index.js';
 export { patterns as modulePatterns } from './module/patterns/index.js';
 export { patterns as uiLibraryPatterns } from './ui-library/patterns/index.js';
-export * from './js-libraries/patterns/index.js';
+export { patterns as datesPatterns } from './dates/patterns/index.js';
 export { patterns as stylingLibraryPatterns } from './styling-libraries/patterns/index.js';
 export { patterns as transpilerPatterns } from './transpiler/patterns/index.js';
