--- packages/analyzer/src/features/js-libraries/patterns/index.ts
@@ -1,5 +1,0 @@
-export { patterns as datePatterns } from './date/index.js';
-export { patterns as httpClientPatterns } from './http-client/index.js';
-export { patterns as translationsPatterns } from './translations/index.js';
-export { patterns as stateManagementPatterns } from './state-management/index.js';
-export { patterns as routerPatterns } from './router/index.js';
