+++ packages/analyzer/src/features/router/index.ts
@@ -1,0 +1,1 @@
+export * from './detect.js';
