+++ packages/analyzer/src/features/router/patterns/index.ts
