+++ packages/analyzer/src/features/router/patterns/react-router.ts
