+++ packages/analyzer/src/features/router/patterns/tanstack-router.ts
