+++ packages/analyzer/src/features/router/patterns/tiny-router.ts
