+++ packages/analyzer/src/features/router/patterns/vue-router.ts
