+++ packages/analyzer/src/features/state-management/index.ts
@@ -1,0 +1,1 @@
+export * from './detect.js';
