+++ packages/analyzer/src/features/state-management/patterns/index.ts
