+++ packages/analyzer/src/features/state-management/patterns/mobx.ts
