+++ packages/analyzer/src/features/state-management/patterns/ngrx.ts
