+++ packages/analyzer/src/features/state-management/patterns/pinia.ts
