+++ packages/analyzer/src/features/state-management/patterns/redux.ts
