+++ packages/analyzer/src/features/translations/index.ts
@@ -1,0 +1,1 @@
+export * from './detect.js';
