+++ packages/analyzer/src/features/translations/patterns/i18next.ts
