+++ packages/analyzer/src/features/translations/patterns/index.ts
