+++ packages/analyzer/src/features/translations/patterns/react-intl.ts
