+++ packages/analyzer/src/features/translations/patterns/vuei18n.ts
